Major Update
Remove redundant code files and file directories
Revise majority of the backend to align with the major changes
In views folder the four roles are removed and replaces with 2 folders, users/admin.
Added additional security features.
Added migration files to accomodate security features.
Created .md documentation for review.
Created .env.production.example for production/deployment.
Change roles in the backend controllers.
Revised web route.
Added securityheader middleware to prevent unknown content to be injected to the system
Added useractivity middleware to handle backend request and log user's last activity.
Added apiresponsetrait to detect a request origin and submit a proper response format.
Added filevalidationtrait to validate submitted files and create a randomize for it in the backend.
Added a LoginAttempt feature that locksout the user for 15 minutes after 5 attempts. each attempt has a 1 minute interval.
Added last_login_at and last_activity_at data at the user table.
Redesign and improve logic of Complete Profile Modal to prompt users to complete profile first.
Revised the footer and nav-bar in the landing page.
Deleted unnecessary files and folder for easier navigation

To-do-Next:
Revise the datepicker in the registration, don't use the calendar-widget and create a custom one.
Test the handlers if it's affect by the major changes
Test the main process of the system. revamp designs/process accordingly.
Observe the system for inconsistencies/bugs due to the major changes.